Add Product scaffold
This adds normal scaffold for Product class, but with Mongo support.
Basically generated using the command from ProductStructure file (see
this commit) and then manually added model file and see (although I was
expecting that only migrations would not be generated).

This actually works pretty much out of the box.

Signed-off-by: Zeljko <zeljko@zwr.fi>