Packed products, scaffold and seed from MH db
Created a PackedProduct as a subclass of Product, and it rocks. It
overrides show_price and works awesome when listed with super-products.
MongoMapper does excellent job setting the list of different object
and setting a correct link on each of them.

Note that the plural form, as created by Rails generators is
product_packeds, but at this point that is not crucial.

Seed file contains those products from MH db which were already
marked, there are some missed, which need to be updated on MH
side first.

Signed-off-by: Zeljko <zeljko@zwr.fi>