Categories contain products
Now categories contain list of product_ids, and custom loaders
for products. Relation is not exactly 1..N, because there is
association field, so that it can be filtered when that product
is in the category. For instance, there can be only one main
category (where is it counted in inventory) and in some cases
there can be views where a product does not appear in a category
(for instance in printed catalogs or price lists, to avoid bloating).

ProductCategory is now renamed Category.

Signed-off-by: Zeljko <zeljko@zwr.fi>