Load instead require seed files
The problem with require is that it will only load if it is not
already loaded, while load essentially executes the file. In the case
of seeds we want to load/execute it, so it seeds the seeds, and if we
do call it again unchanged, it should re-seed. So load it is.

Signed-off-by: Zeljko <zeljko@zwr.fi>