Sass cleanup: @import must be beginning of line
This was actually passing in Chrome, but Minitest was
catching errors in the asset pipe. This is now OK, but
colors have disappeared as now this imported bootstrap
overwrites the other changes. Application.css is placed
last in the asset piping.

Is this import needed at all, given the other one?

Signed-off-by: Zeljko <zeljko@zwr.fi>