Fix 08d2: we need to basename setting
Previous commit didn't work without this. However, if later,
after first generating a template, we switch this off, it seems
that we drag it from the cache.

Would rake tmp:cache:clear help? Or Rails.cache.clear
from rails c? I don't know, I deleted
the whole folder and recloned and now I cannot make
it be as it was.

But now it is nice and short, as we want it. But if we have
many templates, maybe we would need some sorting, or simply
use longer names.

Signed-off-by: Zeljko <zeljko@zwr.fi>